[EI-999] [web] refactor: Add useInstruction hook for providing instructions in linguistic guidance chat
